+++ licensecheck/__main__.py
@@ -1,6 +1,7 @@
 """Entry point for python -m licensecheck."""
+
 from __future__ import annotations
 
-from . import cli
+from licensecheck import cli
 
 cli()