+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2024.1 - 2024/03/16
+
+- code improvements
+- fix issue where onlyLicenses logic caused licensecompat to be `false` when unspecified from the command line
+- move `cli()` to `cli.py`
+- add tests for the main entry point
+
 ## 2024 - 2024/01/27
 
 - code improvements